+++ src/main/java/com/example/Task/manager/repository/TaskRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
-    List<Task> findByUsersId(Long userId);
+public interface TaskRepository extends JpaRepository<Task, Integer> {
+    List<Task> findByUsersId(Integer userId);
 }
 
 /*
